django/requirements.txt
boto3==1.9.189
botocore==1.12.189
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.1
django-allauth==0.39.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-auth==0.9.5
django-storages==1.7.1
django-toolbelt==0.0.1
djangorestframework==3.9.4
djongo==1.2.32
dnspython==1.16.0
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
Pillow==6.1.0
psycopg2==2.8.3
pymongo==3.5.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.2.4
static3==0.7.0
urllib3==1.25.3
whitenoise==4.1.2